requirements.txt
python-dotenv==0.13.0
tda-api==0.3.2
Flask==1.1.2